pyproject.toml
[tool.poetry]
name = "resgen"
version = "0.0.0"
description = "Generate resume based on configuration"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "resgen"}]

[tool.poetry.scripts]
resgen = 'resgen.console.cli:run'

[tool.poetry.dependencies]
python = "^3.10"
fpdf2 = "^2.7.4"
pydantic = "^1.10.9"
pyyaml = "^6.0"
pyyaml-include = "^1.3.1"

[tool.coverage.run]
omit = ["tests/*", "examples/*"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^24.1.1"
mypy = "^1.4.0"
pylint = "^2.17.4"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.1.21"
mkdocs-git-revision-date-localized-plugin = "^1.2.0"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
pattern = """
(?x)                                                        (?# ignore whitespace)
    ^((?P<epoch>\\d+)!)?(?P<base>\\d+(\\.\\d+)*)                   (?# v1.2.3 or v1!2000.1.2)
    ([-._]?((?P<stage>[a-zA-Z]+)[-._]?(?P<revision>\\d+)?))?     (?# b0)
    (\\+(?P<tagged_metadata>.+))?$                               (?# +linux)
 """


[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
